% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_diffusion_model.R
\name{base_diffusion_model}
\alias{base_diffusion_model}
\title{Base Diffusion Model R6 Class}
\usage{
dm <- base_diffusion_model$new(data, model_name="base_dm")

dm$fit(use_bounds=TRUE, transform_pars=FALSE, ...)
}
\arguments{
\item{data}{data.frame or data.table with at least three columns: correctSide, which boundary is the correct answer; response, the response of the subject on that trial; and rt, the response time on the trial}

\item{model_name}{string; arbitrary name for the model}
}
\value{
\code{base_diffusion_model} object
}
\description{
This class serves as a parent object for other diffusion models.
}
\details{
For details for the fit diffusion model method, see \code{\link{fit_diffusion_model}}
}
\section{Fields}{

\describe{
\item{\code{name}}{arbitrary name for model}

\item{\code{data}}{data.frame or data.table with at least three columns: correctSide, which boundary is the correct answer; response, the response of the subject on that trial; and rt, the response time on the trial}

\item{\code{par_names}}{the parameter names for the model}

\item{\code{par_corresponding}}{which underlying diffusion model parameter does this parameter correspond to (important if as_function is used)}

\item{\code{par_values}}{the current parameter values}

\item{\code{start_values}}{starting parameter values (used for optimization)}

\item{\code{obj}}{the objective function used for optimization}

\item{\code{solution}}{information about model fit (returned from the package modelfitr). This attribute is NULL upon initiation, and is created by running the \code{fit} method}
}}

\keyword{internal}
