% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_diffusion_model.R
\docType{data}
\name{diffusion_model}
\alias{diffusion_model}
\title{Drift diffusion model object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
diffusion_model$new(dat, model_name)
}
\arguments{
\item{dat}{data table; contains at least 2 columns: rt - response time for trial, response - upper or lower boundary (1 or 0)}

\item{model_name}{string; name to identify model, default = "ddm"}

\item{include}{character; vector of parameters to include in model. drift rate v, boundary a, and non-decision time t0 are included by default always. Can specify ddm parameters starting point z, drift rate variability sv, non decision time variability st0, starting point variability sz, and wiener diffusion noise s. Also can include collapsing bound parameters: degree of collapse a_prime (weibull only), slope of collapse kappa, time constant of collapse tc.}

\item{depends_on}{named character; if a parameter value depends on a task condition, for example drift rate depends on task difficulty, specify here: c("v" = "task difficulty"). Can specify multiple dependent parameters as a vector}

\item{start_values}{named numeric; to change default starting parameter values, specify as named vector. E.g. c("v"=2, "a"=3) to set drift rate to 2 and boundary to 3}

\item{fixed_pars}{named numeric; to fix a parameter at a specified value, use a named vector as with start_values}

\item{max_time}{numeric; max time to simulate a decision. Lower max time keeps computation times lower, but too low will compromise accuracy}

\item{extra_condition}{character; vector of task condition names. Will calculate first passage times for each condition. Recommended only when comparing a model without depends_on with a model that contains a depends_on parameter.}

\item{bounds_function}{character: "weibull" for weibull function, "hyperbolic_ratio" for hyperbolic ratio function, default = "hyperbolic_ratio"}

\item{objective}{character: "rtdists" to use the rtdists package (pure and extended ddm only, will not work with collapsing bounds), "integral" to use the integral method from Voskuilen et al., 2016, or "chisquare" to use the difference in chisq from actual vs. simulated response times}
}
\value{
definition of base diffusion model object
#'
}
\description{
Drift diffusion model object
}
\keyword{datasets}
