% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evacc_model.R
\name{evacc_model}
\alias{evacc_model}
\title{EvAcc model R6 Class}
\usage{
evacc <- evacc_model$new(dat, stim_var, stim_sep="", stim_dur=.01, stim_interval=.1, model_name="evacc", include=NULL, start_values=NULL, fixed_pars=NULL, bounds=NULL)

evacc$set_objective(objective="chisq")

evacc$fit(use_bounds=TRUE, transform_pars=FALSE, ...)
}
\arguments{
\item{dat}{data table; contains at least 2 columns: rt - response time for trial, response - upper or lower boundary (1 or 0)}

\item{stim_var}{string; name of stimulus column in dat. If two names, first is upper boundary evidence and second is lower boundary evidence. If one variable name, assumes binary evidence with 1=evidence to upper boundary and  0=evidence to lower boundary}

\item{stim_sep}{string; pattern separating timepoint by timepoint evidence within evidence string}

\item{stim_dur}{numeric; stimulus pulse duration (in seconds)}

\item{stim_interval}{numeric; inter-stimulus interval (in seconds)}

\item{stim_pre}{numeric; time before stimulus pulse within stimulus bin}

\item{dt}{numeric; time step to simulate/solve for first passage times}

\item{model_name}{string; name to identify model, default = "ddm"}

\item{include}{character; vector of parameters to include in model. drift rate v, boundary a, and non-decision time t0 are included by default always. Can specify ddm parameters starting point z, drift rate variability sv, non decision time variability st0, starting point variability sz, and wiener diffusion noise s. Also can include collapsing bound parameters: degree of collapse aprime (weibull only), slope of collapse kappa, time constant of collapse tc.}

\item{start_values}{named numeric; to change default starting parameter values, specify as named vector. E.g. c("v"=2, "a"=3) to set drift rate to 2 and boundary to 3}

\item{fixed_pars}{named numeric; to fix a parameter at a specified value, use a named vector as with start_values}

\item{bounds}{string: NOT IMPLEMENTED}
}
\value{
definition of evacc model object
}
\description{
R6 Class that defines an EvAcc model to be applied to a set of behavioral data.
}
\details{
For details regarding other methods, see:
\itemize{
\item{evacc$set_objective: \code{\link{set_evacc_objective}}}
\item{evacc$fit: \code{\link{fit_diffusion_model}}}
}
}
