% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpt_to_sim.R
\name{fpt_to_sim}
\alias{fpt_to_sim}
\title{Get simulated responses and rts from first passage time density}
\usage{
fpt_to_sim(fpt, dt = 0.01, n = 10000)
}
\arguments{
\item{fpt}{matrix; 2 X time points matrix with first passage time for upper and lower boundary crossings}

\item{dt}{numeric; size of time bins used to calculate first passage time density, default = .01}

\item{n}{integer; number of trials to simulate}
}
\value{
data.table with two columns: response and rt
}
\description{
Get simulated responses and rts from first passage time density
}
