% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rt_quantiles.R
\name{get_rt_quantiles}
\alias{get_rt_quantiles}
\title{Get response time quantiles}
\usage{
get_rt_quantiles(
  dat,
  qs = seq(0.1, 0.9, 0.2),
  rt_var = "rt",
  conditions = NULL
)
}
\arguments{
\item{dat}{data.table; choice and response time data}

\item{qs}{numeric; vector of quantiles}

\item{rt_var}{character; name of rt variable in data table}

\item{conditions}{character vector; vector of columns indicating the conditions in which to calcluate rt quantiles}
}
\value{
list with data.table with response time quantiles and vector of rt quantiles
}
\description{
Get response time quantiles
}
