% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_stimulus}
\alias{get_stimulus}
\title{Get pulse stimulus}
\usage{
get_stimulus(sequence, intensity = 1, dur = 0.01, isi = 0.1, dt = 0.001)
}
\arguments{
\item{sequence}{string; string of 0s, and 1s of stimulus values (0 is evidence to lower, 1 is evidence to upper)}

\item{intensity}{numeric; intensity of each pulse, default = 1}

\item{dur}{numeric; duration of pulse, default = .01}

\item{isi}{numeric; inter-pulse interval, default = .1}

\item{dt}{numeric; timestep for simulation, default = .001}
}
\value{
stimulus train
}
\description{
get full pulse stimulus
}
