% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pulse_fp_fpt}
\alias{pulse_fp_fpt}
\title{Get first passage time distribution of pulse diffusion model by simulating probability mass}
\usage{
pulse_fp_fpt(
  stimulus,
  v,
  a,
  t0,
  z = 0.5,
  dc = 0,
  sv = 0,
  st0 = 0,
  sz = 0,
  lambda = 0,
  aprime = 0,
  kappa = 0,
  tc = 0.25,
  uslope = 0,
  umag = 0,
  udelay = 0,
  s = 1,
  v_scale = 1,
  dt = 0.001,
  xbins = 200L,
  bounds = 0L,
  urgency = 0L
)
}
\arguments{
\item{stimulus}{matrix; stimulus to simulate (row 1 is evidence to upper boundary, row 2 to lower boundary)}

\item{v}{numeric; drift rate}

\item{a}{numeric; initial boundary}

\item{t0}{numeric; non-decision time}

\item{z}{numeric; starting point, 0 < z < 1, default = .5}

\item{dc}{numeric; drift criterion, the zero point of the drift rate (the drift rate v = v + dc); default = 0}

\item{sv}{numeric; standard deviation of variability in drift rate, sv >= 0, default = 0}

\item{st0}{numeric; variability in non-decision time. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric; variability in starting point (percentage of boundary a). Normal with mean=z and sd = a*sz. 0 < sz < 1, default = 0}

\item{lambda}{numeric; O-U process slope}

\item{aprime}{numeric; degree of collapse, default = 0}

\item{kappa}{numeric; slope of collapse, default = 0}

\item{tc}{numeric; time constant of collapse, default = .25}

\item{uslope}{numeric; urgency scaling factor, default = 0;}

\item{umag}{numeric; urgency magnitude, default = 0;}

\item{udelay}{numeric; urgency delay, default = 0;}

\item{s}{numeric; standard deviation in wiener diffusion noise}

\item{v_scale}{numeric; scale for the drift rate. drift rate v and variability sv are multiplied by this number}

\item{dt}{numeric; time step of simulation, default = .001}

\item{xbins}{numeric; number of evidence bins, default = 100}

\item{bounds}{int: 0 for fixed, 1 for hyperbolic ratio collapsing bounds, 2 for weibull collapsing bounds, 3 for linear}

\item{urgency}{int; 0 for none, 1 for linear, 2 for logistic}
}
\value{
data frame with three columns: response (1 for upper boundary, 0 for lower), response time, and evidence
}
\description{
Get first passage time distribution of pulse diffusion model by simulating probability mass
}
