% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_diffusion_model.R
\docType{data}
\name{pulse_model}
\alias{pulse_model}
\title{Pulse diffusion model object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
pulse_model$new(dat, model_name)
}
\arguments{
\item{dat}{data table; contains at least 2 columns: rt - response time for trial, response - upper or lower boundary (1 or 0)}

\item{model_name}{string; name to identify model, default = "ddm"}

\item{include}{character; vector of parameters to include in model. drift rate v, boundary a, and non-decision time t0 are included by default always. Can specify ddm parameters starting point z, drift rate variability sv, non decision time variability st0, starting point variability sz, and wiener diffusion noise s. Also can include collapsing bound parameters: degree of collapse a_prime (weibull only), slope of collapse kappa, time constant of collapse tc.}

\item{depends_on}{named character; if a parameter value depends on a task condition, for example drift rate depends on task difficulty, specify here: c("v" = "task difficulty"). Can specify multiple dependent parameters as a vector}

\item{start_values}{named numeric; to change default starting parameter values, specify as named vector. E.g. c("v"=2, "a"=3) to set drift rate to 2 and boundary to 3}

\item{fixed_pars}{named numeric; to fix a parameter at a specified value, use a named vector as with start_values}

\item{extra_condition}{character; vector of task condition names. Will calculate first passage times for each condition. Recommended only when comparing a model without depends_on with a model that contains a depends_on parameter.}

\item{use_weibull_bound}{logical: if T, use weibull function for collapsing bounds. Default = F}
}
\value{
definition of base diffusion model object
#'
}
\description{
Pulse diffusion model object
}
\keyword{datasets}
