% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pulse_nll}
\alias{pulse_nll}
\title{Get pulse model negative log likelihood for a set of trials}
\usage{
pulse_nll(
  choices,
  rt,
  stim_seq,
  v,
  a,
  t0,
  z = 0L,
  sv = 0L,
  st0 = 0L,
  sz = 0L,
  s = 0L,
  lambda = 0L,
  aprime = 0L,
  kappa = 0L,
  tc = 0L,
  check_pars = TRUE,
  dt = 0.002,
  dx = 0.05,
  v_scale = 100,
  use_weibull_bound = FALSE,
  dur = 0.01,
  isi = 0.1,
  n_threads = 1L
)
}
\arguments{
\item{rt}{numeric; vector of response times}

\item{stim_seq}{vector of strings; vector of stimulus sequence for all trials, each element should be a string of 0s and 1s (0 for evidence to lower, 1 for evidence to upper)}

\item{v}{numeric; drift rate, either single value or vector for each trial}

\item{a}{numeric; initial boundary, either single value or vector for each trial}

\item{t0}{numeric; non-decision time, either single value or vector for each trial}

\item{z}{numeric; starting point, , either single value or vector for each trial, 0 < z < 1, default = .5}

\item{sv}{numeric; standard deviation of variability in drift rate, either single value or vector for each trial, sv >= 0, default = 0}

\item{st0}{numeric; variability in non-decision time, either single value or vector for each trial. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric; variability in starting point, either single value or vector for each trial. Uniform from [z-sz/2, z+sz/2], 0 < sz < z, default = 0}

\item{s}{numeric; standard deviation in wiener diffusion noise, either single value or vector for each trial, default = 1}

\item{lambda}{numeric; O-U process slope, either single value or vector for each trial}

\item{aprime}{numeric; degree of collapse, either single value or vector for each trial, default = 0}

\item{kappa}{numeric; slope of collapse, either single value or vector for each trial, default = 0}

\item{tc}{numeric; time constant of collapse, either single value or vector for each trial, default = .25}

\item{check_pars}{logical; if True, check that parameters are vectors of the same length as choices and rts. Must be true if providing scalar parameters. default = true}

\item{dt}{numeric; time step of simulation, default = .002}

\item{dx}{numeric; size of evidence bins, default = .05}

\item{v_scale}{numeric; scale drift rate to be similar to boundary separation a, default = 100}

\item{use_weibull_bound}{logical; if True, use weibull function for collapsing bounds, if False, use hyperbolic ratio function}

\item{dur}{numeric; duration of stimulus}

\item{isi}{numeric; interstimulus interval}

\item{n_threads}{int; number of threads (trials) to run in parallel}

\item{choice}{integer; vector of decisions, 0 for lower boundary, 1 for upper}
}
\value{
negative log likelihood of all choices and rts given pulse model parameters
}
\description{
Get pulse model negative log likelihood for a set of trials
}
