% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pulse_pmass_fpt}
\alias{pulse_pmass_fpt}
\title{Get first passage time distribution of pulse diffusion model by simulating probability mass}
\usage{
pulse_pmass_fpt(stimulus, v, a, t0, z = 0, sv = 0, st0 = 0, sz = 0,
  s = 1, lambda = 0, a_prime = 0, kappa = 0, tc = 0.25,
  dt = 0.002, dx = 0.05, v_scale = 10, use_weibull_bound = FALSE)
}
\arguments{
\item{stimulus}{integer vector; stimulus to simulate (1 for evidence to upper boundary, -1 to lower boundary, 0 for no evidence)}

\item{v}{numeric; drift rate}

\item{a}{numeric; initial boundary}

\item{t0}{numeric; non-decision time}

\item{z}{numeric; starting point, 0 < z < 1, default = .5}

\item{sv}{numeric; standard deviation of variability in drift rate, sv >= 0, default = 0}

\item{st0}{numeric; variability in non-decision time. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric; variability in starting point. Uniform from [z-sz/2, z+sz/2], 0 < sz < z, default = 0}

\item{s}{numeric; standard deviation in wiener diffusion noise, default = 1}

\item{lambda}{numeric; O-U process slope}

\item{a_prime}{numeric; degree of collapse, default = 0}

\item{kappa}{numeric; slope of collapse, default = 0}

\item{tc}{numeric; time constant of collapse, default = .25}

\item{dt}{numeric; time step of simulation, default = .002}

\item{dx}{numeric; size of evidence bins, default = .05}

\item{v_scale}{numeric; scale drift rate to be similar to boundary separation a, default = 10}

\item{use_weibull_bound}{logical; if True, use weibull function for collapsing bounds, if False, use hyperbolic ratio function}
}
\value{
data frame with three columns: response (1 for upper boundary, 0 for lower), response time, and evidence
}
\description{
Get first passage time distribution of pulse diffusion model by simulating probability mass
}
