% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pulse_predict}
\alias{pulse_predict}
\title{Get predicted behavior from pulse model}
\usage{
pulse_predict(
  n,
  stimuli,
  v,
  a,
  t0,
  z = 0L,
  dc = 0L,
  sv = 0L,
  st0 = 0L,
  sz = 0L,
  s = 0L,
  lambda = 0L,
  aprime = 0L,
  kappa = 0L,
  tc = 0L,
  check_pars = TRUE,
  dt = 0.001,
  dx = 0.05,
  bounds = 0L,
  n_threads = 1L
)
}
\arguments{
\item{n}{int; number of predicted samples to take per stimulus}

\item{stimuli}{list; list of stimulus matrices}

\item{v}{numeric; drift rate, either single value or vector for each trial}

\item{a}{numeric; initial boundary, either single value or vector for each trial}

\item{t0}{numeric; non-decision time, either single value or vector for each trial}

\item{z}{numeric; starting point, , either single value or vector for each trial, 0 < z < 1, default = .5}

\item{dc}{numeric; drift criterion, the zero point of the drift rate (the drift rate v = v + dc); default = 0}

\item{sv}{numeric; standard deviation of variability in drift rate, either single value or vector for each trial, sv >= 0, default = 0}

\item{st0}{numeric; variability in non-decision time, either single value or vector for each trial. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric; variability in starting point, either single value or vector for each trial. Uniform from [z-sz/2, z+sz/2], 0 < sz < z, default = 0}

\item{s}{numeric; standard deviation in wiener diffusion noise, either single value or vector for each trial, default = 1}

\item{lambda}{numeric; O-U process slope, either single value or vector for each trial}

\item{aprime}{numeric; degree of collapse, either single value or vector for each trial, default = 0}

\item{kappa}{numeric; slope of collapse, either single value or vector for each trial, default = 0}

\item{tc}{numeric; time constant of collapse, either single value or vector for each trial, default = .25}

\item{check_pars}{logical; if True, check that parameters are vectors of the same length as choices and rts. Must be true if providing scalar parameters. default = true}

\item{dt}{numeric; time step of simulation, default = .001}

\item{bounds}{int: 0 for fixed, 1 for hyperbolic ratio collapsing bounds, 2 for weibull collapsing bounds}

\item{n_threads}{int; number of threads (trials) to run in parallel}
}
\value{
data frame with two columns: response (1 for upper boundary, 0 for lower), response time
}
\description{
Get predicted behavior from pulse model
}
