% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_stimulus.R
\name{pulse_sequence}
\alias{pulse_sequence}
\title{Generate pulse sequence}
\usage{
pulse_sequence(
  n = 1,
  bins = 25,
  dist = "bernoulli",
  p_right = 0.5,
  p_left = NA,
  add_1 = FALSE,
  return_train = FALSE,
  ...
)
}
\arguments{
\item{n}{integer; number of stimuli (or trials)}

\item{bins}{integer; number of stimulus bins}

\item{dist}{string; distribution (or method) used to create stimuli. Either "bernoulli" or "poisson". If "bernoulli" then p_left = 1 - p_right}

\item{p_right}{numeric; probability of right pulse in a given bin}

\item{p_left}{numeric; probability of left pulse in a given bin. if NA or dist = "bernoulli", p_left = 1 - p_right}

\item{add_1}{logical; if TRUE, add pulse at the beginning of each sequence (e.g. if trial starts with pulse on both sides)}

\item{return_train}{logical; if TRUE, returns stimulus matrix (result from pulse_stimulus). if FALSE, returns character matrix. Default = FALSE}

\item{...}{arguments passed to pulse_stimulus (only used if return_train = TRUE)}
}
\value{
stimulus matrix or character matrix
}
\description{
generate new pulse sequence
}
