% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_stimulus.R
\name{pulse_stimulus}
\alias{pulse_stimulus}
\title{Get pulse stimulus}
\usage{
pulse_stimulus(
  up_sequence,
  down_sequence = NULL,
  pattern = "",
  dur = 0.01,
  isi = 0.1,
  pre_stim = 0,
  dt = 0.001
)
}
\arguments{
\item{up_sequence}{string; string timepoint by timepoint evidence to upper boundary}

\item{down_sequence}{string; string timepoint by timepoint evidence to upper boundary (optional)}

\item{pattern}{string; pattern to separate timepoints. Default = ""}

\item{dur}{numeric; duration of pulse, default = .01}

\item{isi}{numeric; inter-pulse interval, default = .1}

\item{pre_stim}{numeric; time before pulse within stimulus bin. default = 0}

\item{dt}{numeric; timestep for simulation, default = .001}
}
\value{
stimulus train
}
\description{
get full pulse stimulus
}
