% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pulse_trial_stimulus}
\alias{pulse_trial_stimulus}
\title{Get pulse stimulus}
\usage{
pulse_trial_stimulus(stim_seq, dur = 0.01, isi = 0.1, pre_stim = 0, dt = 0.001)
}
\arguments{
\item{stim_seq}{numeric vector; vecotr of timepoint by timepoint evidence}

\item{dur}{numeric; duration of pulse, default = .01}

\item{isi}{numeric; inter-pulse interval, default = .1}

\item{pre_stim}{numeric; time before pulse within stimulus bin}

\item{dt}{numeric; timestep for simulation, default = .001}
}
\value{
stimulus train vector
}
\description{
get full pulse stimulus
}
