% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{qmpe}
\alias{qmpe}
\title{qmpe objective function}
\usage{
qmpe(rts, rt_qs, p_q, n_rt, min_p = 1e-10)
}
\arguments{
\item{rts}{list of simulated response times for different conditions}

\item{rt_qs}{matrix of response time quantiles for each condition (condition x rt quantile value)}

\item{p_q}{vector of probabilities corresponding to response time quantiles}

\item{n_rt}{vector with the number of empirical response times for each condition}

\item{min_p}{numeric; minimum probability in bin}
}
\value{
numeric scalar, qmpe value
}
\description{
qmpe objective function
}
