% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{quantile_chisquare}
\alias{quantile_chisquare}
\title{chisquare objective function}
\usage{
quantile_chisquare(rts, rt_qs, p_q, n_rt)
}
\arguments{
\item{rts}{list of simulated response times for different conditions}

\item{rt_qs}{matrix of response time quantiles for each condition (condition x rt quantile value)}

\item{p_q}{vector of probabilities corresponding to response time quantiles}

\item{n_rt}{vector with the number of empirical response times for each condition}
}
\value{
numeric scalar, chisquare value
}
\description{
chisquare objective function
}
