% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddm-package.R
\docType{package}
\name{rddm-package}
\alias{rddm}
\alias{rddm-package}
\title{rddm: Fit diffusion models to experimental data}
\description{
Pure DDM and extended DDM (Bogacz et al., 2006) using rtdist package Collapsing bounds (Voskuilen et al., 2016) Pulse diffusion model (see Brunton et al., 2013)
}
\keyword{internal}
