% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_diffusion_model.R
\name{set_ddm_objective}
\alias{set_ddm_objective}
\title{set diffusion model objective function (for internal use)}
\usage{
model$set_objective(objective)
}
\arguments{
\item{objective}{string; the objective function to be used; either "rtdists", "integral", "chisq", or "qmpe"}
}
\value{
modifies the field \code{obj}
}
\description{
Change the objective function used to fit the diffusion model. Options include:
\describe{
\item{rtdists}{use the fast-dm method implemented in the rtdists package}
\item{integral}{use the integral equation method from Smith, 2000 to obtain the first passage time distribution}
\item{chisq}{simulate the first passage time distribution using the Euler-Maruyama method, and compare simulated and observed distrubtions using the chisquare statistic}
}
}
