% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evacc_model.R
\name{set_evacc_objective}
\alias{set_evacc_objective}
\title{set evacc model objective function (for internal use)}
\usage{
model$set_objective(objective)
}
\arguments{
\item{objective}{string; the objective function to be used; only "chisq" implemented}
}
\value{
modifies the field \code{obj}
}
\description{
Change the objective function used to fit the evacc model. Options include:
\describe{
\item{chisq}{simulate the first passage time distribution using the Euler-Maruyama method, and compare simulated and observed distrubtions using the chisquare statistic}
}
}
