% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_ddm_vec}
\alias{sim_ddm_vec}
\title{Simulate drift diffusion model with fixed or collapsing boundary}
\usage{
sim_ddm_vec(
  v,
  a,
  t0,
  z = 0L,
  dc = 0L,
  sv = 0L,
  st0 = 0L,
  sz = 0L,
  aprime = 0L,
  kappa = 0L,
  tc = 0L,
  uslope = 0L,
  udelay = 0L,
  umag = 0L,
  s = 0L,
  dt = 0.001,
  max_time = 10,
  bounds = 0L,
  urgency = 0L,
  check_pars = TRUE,
  n_threads = 1L,
  return_accu = FALSE,
  seed = -1L
)
}
\arguments{
\item{v}{vector; drift rate}

\item{a}{vector; initial boundary}

\item{t0}{vector; non-decision time}

\item{z}{vector; starting point, 0 < z < 1, default = .5}

\item{dc}{vector; drift criterion, the zero point of the drift rate (the drift rate v = v + dc); default = 0}

\item{sv}{vector; standard deviation of variability in drift rate, sv >= 0, default = 0}

\item{st0}{vector; variability in non-decision time. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{vector; variability in starting point (percentage of boundary a). Uniform from [z-sz/2, z+sz/2], 0 < sz < z, default = 0}

\item{aprime}{vector; degree of collapse, default = 0}

\item{kappa}{vector; slope of collapse, default = 0}

\item{tc}{vector; time constant of collapse, default = .25}

\item{uslope}{numeric; urgency scaling factor, default = 0;}

\item{udelay}{numeric; urgency delay, default = 0;}

\item{umag}{numeric; urgency magnitude, default = 0;}

\item{s}{numeric; standard deviation in wiener diffusion noise, default = 1}

\item{dt}{numeric; time step of simulation, default = .001}

\item{max_time}{numeric; max time of simulation, default = 10}

\item{bounds}{int: 0 for fixed, 1 for hyperbolic ratio collapsing bounds, 2 for weibull collapsing bounds}

\item{urgency}{int: 0 for none, 1 for linear, 2 for logistic}

\item{check_pars}{bool; if True (default) check parameter vector lengths and default values}

\item{n_threads}{integer; number of threads to run in parallel, default = 1}

\item{return_accu}{bool; if True, return full trajectory of accumulators}

\item{seed}{integer; set random seed}
}
\value{
data frame with two columns: response (1 for upper boundary, 0 for lower), and response time
}
\description{
Simulate drift diffusion model with fixed or collapsing boundary
}
