% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_evacc}
\alias{sim_evacc}
\title{Simulate EvAcc (2-accumulator) model with fixed or collapsing boundary}
\usage{
sim_evacc(
  n,
  stimuli,
  a,
  t0,
  z = 0,
  dc = 0,
  sv = 0,
  st0 = 0,
  sz = 0,
  s = 1,
  lambda = 0,
  mi = 0,
  sv2 = 0,
  s2 = 1,
  lambda2 = 1,
  mi2 = 0,
  aprime = 0,
  kappa = 0,
  tc = 0.25,
  v = 1,
  accumulator_gain = FALSE,
  two_layer = FALSE,
  scalar_stimulus_noise = FALSE,
  scalar_diffusion_noise = FALSE,
  v_scale = 1,
  dt = 0.001,
  bounds = 0L,
  n_threads = 1L,
  return_accu = FALSE
)
}
\arguments{
\item{n}{integer; number of simulations for each stimulus}

\item{a}{numeric; initial boundary}

\item{t0}{numeric; non-decision time}

\item{z}{numeric; boundary separation bias, left boundary = a - z, right = a + z}

\item{dc}{numeric; drift criterion, left = v + dc, right = v - dc}

\item{sv}{numeric; standard deviation of variability in drift rate, sv >= 0, default = 0}

\item{st0}{numeric; variability in non-decision time. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric; variability in starting bias (drawn from gaussian), sz > 0, default = 0}

\item{s}{numeric; diffusion noise}

\item{lambda}{numeric; accumulator recurrent weight}

\item{mi}{numeric; mutual inhibition}

\item{sv2}{numeric; sampling noise for second layer}

\item{s2}{numeric; diffusion noise for layer 2}

\item{lambda2}{numeric; second layer accumulator recurrent weight}

\item{mi2}{numeric; mutual inhibition in second layer. Set mi=0 and mi2 > 0 to implement static sample model from Scott et al., 2015}

\item{aprime}{numeric; degree of collapse, default = 0}

\item{kappa}{numeric; slope of collapse, default = 1}

\item{tc}{numeric; time constant of collapse, default = .25}

\item{v}{numeric; drift rate. Default = 1, recommended to leave this parameter fixed}

\item{two_layer}{bool; if true, use two layer accumulator model}

\item{scalar_stimulus_noise}{bool; if true, drift rate variability scales with accumulator value (see Koay et al., 2020)}

\item{scalar_diffusion_noise}{bool; if true, diffusion noise scales with accumulator value}

\item{v_scale}{numeric; scale for the drift rate. drift rate v and variability sv and criterion dc are multiplied by this number}

\item{dt}{numeric; time step of simulation, default = .001}

\item{bounds}{int: 0 for fixed, 1 for hyperbolic ratio collapsing bounds, 2 for weibull collapsing bounds}

\item{n_threads}{integer; number of threads to run in parallel, default = 1}

\item{return_accu}{bool; if True, return full trajectory of accumulators}

\item{stimulus}{array; stimulus to simulate, 1 or 2 rows X timepoints x number of stimuli. If 2 rows, row 1 is evidence to upper boundary, row 2 to lower boundary.}

\item{accumulator_gain;}{if true, gain of first accumulator layer scales with value}
}
\value{
List containing 1) data frame with three columns: response (1 for right, 0 for left), response time, and evidence and 2) matrix with full accumulator trajectories
}
\description{
Simulate EvAcc (2-accumulator) model with fixed or collapsing boundary
}
