% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pulse}
\alias{sim_pulse}
\title{Simulate drift diffusion model with fixed or collapsing boundary}
\usage{
sim_pulse(
  n,
  up_stimulus,
  down_stimulus,
  v,
  a,
  t0,
  z = 0.5,
  d = 0,
  sv = 0,
  st0 = 0,
  sz = 0,
  s = 1,
  lambda = 0,
  aprime = 0,
  kappa = 0,
  tc = 0.25,
  dt = 0.001,
  v_scale = 100,
  use_weibull_bound = FALSE,
  n_threads = 1L
)
}
\arguments{
\item{n}{integer; number of decisions to simulate}

\item{up_stimulus}{float vector; stimulus to the positive boundary}

\item{down_stimulus}{float vector; stimulus to the negative boundary}

\item{v}{numeric; drift rate}

\item{a}{numeric; initial boundary}

\item{t0}{numeric; non-decision time}

\item{z}{numeric; starting point, 0 < z < 1, default = .5}

\item{d}{numeric; drift criterion, upper drift = v, lower drift = v-d}

\item{sv}{numeric; standard deviation of variability in drift rate, sv >= 0, default = 0}

\item{st0}{numeric; variability in non-decision time. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric; variability in starting point. Uniform from [z-sz/2, z+sz/2], 0 < sz < z, default = 0}

\item{s}{numeric; standard deviation in wiener diffusion noise, default = 1}

\item{lambda}{numeric; O-U process slope}

\item{aprime}{numeric; degree of collapse, default = 0}

\item{kappa}{numeric; slope of collapse, default = 1}

\item{tc}{numeric; time constant of collapse, default = .25}

\item{dt}{numeric; time step of simulation, default = .001}

\item{v_scale}{numeric; scale drift rate to be similar to boundary separation a, default = 100}

\item{use_weibull_bound}{logical; if True, use weibull function for collapsing bounds, if False, use hyperbolic ratio function}

\item{n_threads}{integer; number of threads to run in parallel, default = 1}
}
\value{
data frame with three columns: response (1 for upper boundary, 0 for lower), response time, and evidence
}
\description{
Simulate drift diffusion model with fixed or collapsing boundary
}
