% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pulse_vec}
\alias{sim_pulse_vec}
\title{Simulate drift diffusion model with fixed or collapsing boundary}
\usage{
sim_pulse_vec(
  n,
  stimuli,
  v,
  a,
  t0,
  z = 0L,
  dc = 0L,
  sv = 0L,
  st0 = 0L,
  sz = 0L,
  s = 0L,
  lambda = 0L,
  aprime = 0L,
  kappa = 0L,
  tc = 0L,
  v_scale = 1,
  dt = 0.001,
  bounds = 0L,
  check_pars = TRUE,
  n_threads = 1L
)
}
\arguments{
\item{n}{integer; number of decisions to simulate}

\item{v}{numeric vector; drift rate}

\item{a}{numeric vector; initial boundary}

\item{t0}{numeric vector; non-decision time}

\item{z}{numeric vector; starting point, 0 < z < 1, default = .5}

\item{dc}{numeric vector; drift criterion, upper drift = v, lower drift = v-d}

\item{sv}{numeric vector; standard deviation of variability in drift rate, sv >= 0, default = 0}

\item{st0}{numeric vector; variability in non-decision time. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric vector; variability in starting point. Uniform from [z-sz/2, z+sz/2], 0 < sz < z, default = 0}

\item{s}{numeric vector; standard deviation in wiener diffusion noise, default = 1}

\item{lambda}{numeric vector; O-U process slope}

\item{aprime}{numeric vector; degree of collapse, default = 0}

\item{kappa}{numeric vector; slope of collapse, default = 1}

\item{tc}{numeric vector; time constant of collapse, default = .25}

\item{v_scale}{numeric; scale for the drift rate. drift rate v and variability sv are multiplied by this number}

\item{dt}{numeric; time step of simulation, default = .001}

\item{bounds}{int: 0 for fixed, 1 for hyperbolic ratio collapsing bounds, 2 for weibull collapsing bounds}

\item{check_pars}{bool; if True (default) check parameter vector lengths and default values}

\item{n_threads}{integer; number of threads to run in parallel, default = 1}

\item{stimulus}{matrix; stimulus to simulate, 1 or 2 rows X timepoints. If 2 rows, row 1 is evidence to upper boundary, row 2 to lower boundary.}
}
\value{
List containing 1) data frame with three columns: response (1 for upper boundary, 0 for lower), response time, and evidence and 2) matrix with full accumulator trajectories
}
\description{
Simulate drift diffusion model with fixed or collapsing boundary
}
