% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_to_fpt.R
\name{sim_to_fpt}
\alias{sim_to_fpt}
\title{Get first passage time density from simulated diffusion model data}
\usage{
sim_to_fpt(sim_data, sim_dt = 0.001, fpt_dt = 0.01, max_time = 10)
}
\arguments{
\item{sim_data}{numeric; drift rate}

\item{sim_dt}{numeric; time step of simulation, default = .001}

\item{fpt_dt}{numeric; size of time bins to calculate first passage time, default = .01}

\item{max_time}{numeriic; max time to calculate probability of first passage times}
}
\value{
matrix with two columns: first column is p(upper), second is p(lower)
}
\description{
Get first passage time density from simulated diffusion model data
}
