% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_diffusion_model.R
\name{simulate_pulse_model}
\alias{simulate_pulse_model}
\title{simulate pulse model  (for internal use)}
\usage{
model$simulate(n, stimuli, par_values, par_names=NULL, ...)
}
\arguments{
\item{n}{integer; number of decisions to simulate for each condition. If the number of conditions is equal to the length of the data, e.g. if using as_function with a continuous predictor, ignores \code{n} and simulates one decision per condition}

\item{stimuli}{array; 2 x timepoints x trials array of stimuli (can generate using \code{pulse_stimulus} or \code{pulse_sequence})}

\item{par_values}{numeric vector; vector of parameters. Must be named vector or used with par_names}

\item{par_names}{character vector; vector of parameter names}

\item{...}{additional arguments passed to \code{sim_pulse}}
}
\value{
data.table with simulation conditions, decision (upper or lower boundary) and response time
}
\description{
simulate pulse DDM with given model parameters..
This function is only intended for use with a pulse model object,
and should not be called directly outside of the pulse model class.
Please use \code{sim_pulse} as a standalone function.
}
\keyword{internal}
