% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\alias{logistic}
\alias{inv_logistic}
\title{Parameter Trasnform functions}
\usage{
logistic(x, lower, upper)

inv_logistic(x, lower, upper)
}
\arguments{
\item{x}{numeric; vector of parameter values}

\item{lower}{numeric; parameter lower bound}

\item{upper}{numeric; parameter upper bound}
}
\value{
transformed parameter vector
}
\description{
Parameter Transform Functions
}
