% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{urgency}
\alias{urgency}
\alias{linear_urgency}
\alias{logistic_urgency}
\title{Urgency functions}
\usage{
linear_urgency(t, uslope = 0, udelay = 0, umag = 0)

logistic_urgency(t, uslope = 0, udelay = 0, umag = 0)
}
\arguments{
\item{t}{vector; time points to evaluate boundary}

\item{uslope;}{slope of urgency signal}

\item{udelay;}{delay to start of rising urgency}

\item{umag;}{magnitude of urgency signal (only for logistic)}
}
\value{
column vector with urgency at time t,
}
\description{
evaluate diffusion model urgency signals (linear and logistic)
}
