% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zrand}
\alias{zrand}
\alias{zrandn}
\alias{zrandseed}
\title{random normal}
\usage{
zrandn(n)

zrandseed(s)
}
\arguments{
\item{n}{int; number of random draws}
}
\value{
random numbers
}
\description{
random normal
}
