% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_diffusion_model.R
\docType{data}
\name{base_diffusion_model}
\alias{base_diffusion_model}
\title{Base diffusion model object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
base_diffusion_model$new(dat, model_name)
}
\arguments{
\item{dat}{data table; contains at least 2 columns: rt - response time for trial, response - upper or lower boundary (1 or 0)}

\item{model_name}{string (optional); name to identify model}
}
\value{
definition of base diffusion model object
}
\description{
Base diffusion model object
}
\keyword{datasets}
