% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_diffusion_model.R
\name{diffusion_model}
\alias{diffusion_model}
\title{Drift Diffusion Model R6 Class}
\usage{
dm <- diffusion_model$new(dat, model_name="ddm", include=NULL, depends_on=NULL, as_function=NULL, start_values=NULL, fixed_pars=NULL, max_time=10, extra_condition=NULL, bounds=NULL, objective=NULL, ...)

dm$set_objective(objective=NULL)

dm$fit(use_bounds=TRUE, transform_pars=FALSE, ...)

dm$predict(pars=NULL, n=10000, ...)

dm$simulate(n, par_values)
}
\arguments{
\item{dat}{data table; contains at least 2 columns: rt - response time for trial, response - upper or lower boundary (1 or 0)}

\item{model_name}{string; name to identify model, default = "ddm"}

\item{include}{character; vector of parameters to include in model. drift rate v, boundary a, and non-decision time t0 are included by default always. Can specify ddm parameters starting point z, drift rate variability sv, non decision time variability st0, starting point variability sz, and wiener diffusion noise s. Also can include collapsing bound parameters: degree of collapse aprime (weibull only), slope of collapse kappa, time constant of collapse tc.}

\item{depends_on}{named character; if a parameter value depends on a task condition, for example drift rate depends on task difficulty, specify here: c("v" = "task difficulty"). Can specify multiple dependent parameters as a vector}

\item{as_function}{list; list specifying parameters whose value is a function of other parameters}

\item{start_values}{named numeric; to change default starting parameter values, specify as named vector. E.g. c("v"=2, "a"=3) to set drift rate to 2 and boundary to 3}

\item{fixed_pars}{named numeric; to fix a parameter at a specified value, use a named vector as with start_values}

\item{max_time}{numeric; max time to simulate a decision. Lower max time keeps computation times lower, but too low will compromise accuracy}

\item{extra_condition}{character; vector of task condition names. Will calculate first passage times for each condition. Recommended only when comparing a model without depends_on with a model that contains a depends_on parameter.}

\item{bounds}{string: either "fixed" for fixed bounds, or "weibull" or "hyperbolic" for collapsing bounds according to weibull or hyperbolic ratio functions}

\item{objective}{character: "rtdists" to use the rtdists package (pure and extended ddm only, will not work with collapsing bounds), "integral" to use the integral method from Voskuilen et al., 2016, "chisquare" to use the difference in chisq from actual vs. simulated response times, or "qmpe" to use the quantile maximum product estimation method (comparing simulated vs. actual RT distribution) from Brown & Heathcote}

\item{...}{additional arguments passed to the objective function}
}
\value{
\code{base_diffusion_model} object
}
\description{
R6 Class that defines a diffusion model to be applied to a set of behavioral data.
This class implements all of the following models:
\itemize{
\item{pure DDM}
\item{extended DDM}
\item{collapsing bounds using hyperbolic ratio (Voskuilen et al., 2016) or weibull (Hawkins et al., 2015) function}
}
}
\details{
For details regarding other methods, see:
\itemize{
\item{dm$set_objective: \code{\link{set_ddm_objective}}}
\item{dm$fit: \code{\link{fit_diffusion_model}}}
\item{dm$predict: \code{\link{predict_diffusion_model}}}
\item{dm$simulate: \code{\link{simulate_diffusion_model}}}
}
}
\section{Fields}{

\describe{
\item{\code{name}}{arbitrary name for model}

\item{\code{data}}{data.frame or data.table with at least three columns: correctSide, which boundary is the correct answer; response, the response of the subject on that trial; and rt, the response time on the trial}

\item{\code{data_q}}{response times quantile (only used if objective = "chisq")}

\item{\code{par_names}}{the parameter names for the model}

\item{\code{par_corresponding}}{which underlying diffusion model parameter does this parameter correspond to (important if as_function is used)}

\item{\code{par_values}}{the current parameter values}

\item{\code{start_values}}{starting parameter values (used for optimization)}

\item{\code{obj}}{the objective function used for optimization}

\item{\code{solution}}{information about model fit (returned from the package modelfitr). This attribute is NULL upon initiation, and is created by running the \code{fit} method}
}}

