% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_diffusion_model.R
\name{fit_diffusion_model}
\alias{fit_diffusion_model}
\title{fit diffusion model (for internal use)}
\usage{
model$fit(use_bounds=TRUE, transform_pars=FALSE, ...)
}
\arguments{
\item{use_bounds}{logical; if True, perform bounded optimization}

\item{transform_pars}{logical; if True, will use logistic transformation on parameters to ensure they don't violate bounds, even if using an unbounded optimization method}

\item{n_cores}{integer; number of cores to use for parallel processing (for sections using the foreach package)}

\item{...}{additional arguments passed to \code{modelfitr::fit_model}}
}
\value{
\code{base_diffusion_model} with result saved to the field \code{solution}
}
\description{
Fit parameters of the diffusion model.
This function is only intended for use with a diffusion model object,
and should not be called directly outside of the diffusion model class.
}
\keyword{internal}
