% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_pulse}
\alias{sim_pulse}
\title{Simulate pulse diffusion model with fixed or collapsing boundary}
\usage{
sim_pulse(
  n,
  stimuli,
  a,
  t0,
  s,
  z = 0.5,
  dc = 0,
  sv = 0,
  st0 = 0,
  sz = 0,
  lambda = 0,
  aprime = 0,
  kappa = 0,
  tc = 0.25,
  uslope = 0,
  umag = 0,
  udelay = 0,
  v_scale = 1,
  dt = 0.001,
  bounds = 0L,
  urgency = 0L,
  n_threads = 1L,
  return_accu = FALSE
)
}
\arguments{
\item{n}{integer; number of decisions to simulate}

\item{stimuli}{cube; stimuli to simulate, 2 rows X timepoints x trials.}

\item{a}{numeric; initial boundary}

\item{t0}{numeric; non-decision time}

\item{s}{numeric; standard deviation in wiener diffusion noise}

\item{z}{numeric; starting point, 0 < z < 1, default = .5}

\item{dc}{numeric; drift criterion, upper drift = v, lower drift = v-d}

\item{sv}{numeric; standard deviation of variability in drift rate, sv >= 0, default = 0}

\item{st0}{numeric; variability in non-decision time. Uniform from [t0-st0/2, t0+st0/2], 0 < st0 < t0, default = 0}

\item{sz}{numeric; variability in starting point. Uniform from [z-sz/2, z+sz/2], 0 < sz < z, default = 0}

\item{lambda}{numeric; O-U process slope}

\item{aprime}{numeric; degree of collapse, default = 0}

\item{kappa}{numeric; slope of collapse, default = 1}

\item{tc}{numeric; time constant of collapse, default = .25}

\item{uslope}{numeric; urgency scaling factor, default = 0;}

\item{umag}{numeric; urgency magnitude, default = 0;}

\item{udelay}{numeric; urgency delay, default = 0;}

\item{v_scale}{numeric; scale for the drift rate. drift rate v and variability sv are multiplied by this number}

\item{dt}{numeric; time step of simulation, default = .001}

\item{bounds}{int: 0 for fixed, 1 for hyperbolic ratio collapsing bounds, 2 for weibull collapsing bounds, 3 for linear}

\item{urgency}{int: 0 for none, 1 for linear, 2 for logistic}

\item{n_threads}{integer; number of threads to run in parallel, default = 1}

\item{return_accu}{bool; if True, return full trajectory of accumulators}
}
\value{
List containing 1) data frame with three columns: response (1 for upper boundary, 0 for lower), response time, and evidence and 2) matrix with full accumulator trajectories
}
\description{
Simulate pulse diffusion model with fixed or collapsing boundary
}
