% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_diffusion_model.R
\name{simulate_pulse_model}
\alias{simulate_pulse_model}
\title{simulate pulse model  (for internal use)}
\usage{
model$simulate(n, stimuli, par_values, par_names=NULL, ...)
}
\arguments{
\item{n}{integer; number of decisions to simulate per stimulus}

\item{pars}{numeric vector; vector of parameters}

\item{...}{additional arguments passed to \code{sim_pulse}}
}
\value{
data.table with simulation conditions, decision (upper or lower boundary) and response time
}
\description{
simulate pulse DDM with given model parameters..
This function is only intended for use with a pulse model object,
and should not be called directly outside of the pulse model class.
Please use \code{sim_pulse} as a standalone function.
}
\keyword{internal}
