% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{urgency}
\alias{urgency}
\alias{linear_urgency}
\alias{linear_urgency_vec}
\alias{logistic_urgency}
\alias{logistic_urgency_vec}
\title{Urgency functions}
\usage{
linear_urgency(t, uslope = 0, udelay = 0, umag = 0)

linear_urgency_vec(t, uslope = 0L, udelay = 0L, umag = 0L, check_pars = TRUE)

logistic_urgency(t, uslope = 0, udelay = 0, umag = 0)

logistic_urgency_vec(t, uslope = 0L, udelay = 0L, umag = 0L, check_pars = TRUE)
}
\arguments{
\item{t}{vector; time points to evaluate boundary}

\item{uslope;}{slope of urgency signal}

\item{udelay;}{delay to start of rising urgency}

\item{umag;}{magnitude of urgency signal (only for logistic)}
}
\value{
column vector with urgency at time t,
}
\description{
evaluate diffusion model urgency signals (linear and logistic)
}
